// client/src/types/schema.ts
export interface FitnessData {
    id: string;
    patientId?: string;
    date: Date;
    steps?: number;
    waterIntake?: number; // glasses
    sleepHours?: string;
    exerciseMinutes?: number;
    heartRate?: number;
    bloodPressure?: string;
    notes?: string;
}